+ ADDED NAVIGATION, refactoring
refactored the naming so everything is nice and uniform.
The menu now has navigation buttons on the second page to show navigation between screens, and how its implemted in comments.
need to make screen1,2,3 actually named something unique
need to make screen1,2,3 uniform in terms of theme and header.
need to finalise menu.c as an actual menu.
Proof of concept is there though and nav works! yay!